requirements.txt
# Herta's Eye - Python 依赖
# Python 3.10+

# 后端
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# 前端
streamlit>=1.28.0
pandas>=2.0.0
altair>=5.0.0
requests>=2.31.0

# 硬件监控 (Pythonnet)
pythonnet>=3.0.0

# 系统托盘
pystray>=0.19.0
Pillow>=10.0.0
